Refactor error aggregation
