+++ boyar/flow.go
@@ -1,0 +1,36 @@
+package boyar
+
+import (
+	"context"
+	"github.com/orbs-network/boyarin/boyar/config"
+	"github.com/orbs-network/boyarin/strelets"
+	"github.com/orbs-network/boyarin/strelets/adapter"
+	"github.com/orbs-network/scribe/log"
+)
+
+func Flow(ctx context.Context, cfg config.NodeConfiguration, configCache config.BoyarConfigCache, logger log.Logger) error {
+	orchestrator, err := adapter.NewDockerSwarm(cfg.OrchestratorOptions())
+	if err != nil {
+		return err
+	}
+	defer orchestrator.Close()
+
+	s := strelets.NewStrelets(orchestrator)
+	b := NewBoyar(s, cfg, configCache, logger)
+
+	var errors []error
+
+	if err := b.ProvisionVirtualChains(ctx); err != nil {
+		errors = append(errors, err)
+	}
+
+	if err := b.ProvisionHttpAPIEndpoint(ctx); err != nil {
+		errors = append(errors, err)
+	}
+
+	if len(errors) > 0 {
+		return aggregateErrors(errors)
+	}
+
+	return nil
+}
